requirements.txt
# Data
yfinance>=0.2.40
akshare>=1.14.0
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.5.0
PyYAML>=6.0
httpx>=0.27.0
pdfplumber>=0.11.0

# LLM (Commentary drafter, Stage 5)
anthropic>=0.40.0

# Env
python-dotenv>=1.0.0

# Rendering / screenshots (Stage 7)
jinja2>=3.1.0
# playwright is installed separately via `playwright install chromium` after pip
playwright>=1.40.0
